fetch coupon and manage values
The coupon is grabbed with all its data, including meta (where Multicurrency Amounts = main currency amount - are).

The new amount is calculated (converted), if needed.
It all works based on assumption that one Multicurrency Amount is defined, and then Aelia's exchange rate is applied to convert it to any currently chosen currency.

/ In this delivery only the amount is calculated - not yet applied to Cart /